requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
bitsandbytes==0.41.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
diffusers==0.29.2
dill==0.3.7
docstring_parser==0.17.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
flash_attn==2.5.9.post1
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gradio_client==0.2.9
h11==0.14.0
hf-xet==1.1.9
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
latex2mathml==3.78.1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.15
networkx==3.4.2
ninja==1.13.0
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==3.1.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
peft==0.4.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pydantic==1.10.8
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
sentencepiece==0.1.99
sentry-sdk==2.36.0
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.3
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.15.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.40.0.dev0
triton==2.1.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.35.0
wandb==0.21.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
